refactor: remove custom GraphQL type definitions and update integration managers to use postgres-derived types.
