+++ packages/server/utils/getGroupedReactjis.ts
@@ -2,7 +2,7 @@ import ReactjiId from 'parabol-client/shared/gqlIds/ReactjiId'
 import type {ReactjiSource} from '../graphql/public/types/Reactji'
 import type {ReactjiDB} from '../postgres/types'
 
-const getGroupedReactjis = (reactjis: ReactjiDB[], viewerId: string, idPrefix: string) => {
+const getGroupedReactjis = (reactjis: ReactjiDB[], viewerId: string, idPrefix: string | number) => {
   const agg = {} as {[key: string]: ReactjiSource}
   reactjis.forEach((reactji) => {
     const {id, userId} = reactji
